+++ common/src/entry.rs
